+++ requirements.txt
@@ -1,5 +1,6 @@
 matplotlib
 pandoc
+joblib
 pennylane
 qiskit
 qiskit-aer
